pyproject.toml
[tool.poetry]
name = "snow-wise"
version = "0.1.0"
description = ""
authors = ["Shubham Pandey <shubhampcpandey13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-openai = "^0.1.6"
langchain-community = "^0.0.38"
snowflake-sqlalchemy = "^1.5.3"
langchain = "^0.1.20"
streamlit = "^1.34.0"
snowflake-connector-python = "^3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
